add the recalled battery information for Toshiba
Some Toshiba batteries have recently been recalled. Data from
bxinfo.toshiba.com was used to mark these
batteries as recalled in the hal device tree so a warning can
be given to the user to check the battery.
Care has been taken to ensure accuracy, but please don't take
the warning as a message the battery is recalled, nor that the
absence of a warning means the battery is not recalled.
Please use the website to be sure in either case. Thanks.