cleanup and downsize usb music player FDI file
Cleaned up the FDI-file for USB music players. Removed the common properties
from the specific matches and added them to the end of the file to get them
merged to all devices with the key portable_audio_player.type.

This make the file smaller and reduce the need to always add the same keys for
each device (and reduce IMO also the fdi-cache file).

If you add devices to the list which don't need all the common keys, add the
device to the end of the list.