fix dell d420 suspend quirks
According to [1] and [2], the quirks for the Dell D410 laptop in
hal-info break the suspend/resume (because of vbestate_restore). The
attached patch fixes that (and uses the same quirks as s2ram).

I guess, the quirks for D420 should be updated accordingly (at least
s2ram uses the same vbe_post/vbemode_restore quirk for D410 and D420).

Cheers,
Michael