disable polling addon for select USB sticks
> > among the various patches in mandriva package for hal, I have one to
> > deal with some usb keys we have here which doesn't support getting
> > polled at all, causing a lot of kernel error in syslog. Moreover, there
> > isn't really a need to poll these kind of usb devices, since kernel
> > events will notify when it is unplugged from the system.
> >
> > I have some kind of ugly patch (attached) which disable polling for
> > those devices after initial fs detection but it is only effective once
> > usb key is mounted.
> >
> > What do you think of adding a storage.disable_media_check (or maybe
> > disable_polling) property to have this kind of device ?
>
> Wouldn't it work with just an fdi file that sets storage.removable to
> FALSE?

Just tried, it didn't work.

But it works when disabling storage.media_check_enabled.

I've attached a patch for hal-info for two keys we have with this
problem.

I'm wondering if polling is useful at all on usb sticks. But maybe we
are forced to use it since we can't distinguish them from cards reader
or floppy usb drive.