fix dell latitude c800 video quirk
  On a Dell Latitude C800 with no quirks, resume from hibernation
does not restore a useable X session (i.e. a casual observer would
state resume from hibernation fails).  One way to entice the resume
to function correctly is via the quirk vbemode_restore, as included
in the attached patch for hal-info.  (When resuming into tty1, only
selected characters in the lines above the initial prompt are
visibly displayed, but all characters on the initial prompt line
and following are correctly displayed.)

  I was cheeky enough to sort the matches within the Latitude group,
as well as adding a new match.  While the software should do quite
well with these matches in random order, the human updaters may have
less confusion if these remain sorted.  If someone thinks this is
helpful, they may like to also sort Inspiron before Latitude,
etc. so that there should be even less of a risk of the file confusing
its human readers.