add keymap for acer 1360, 1520, 1690, 3020, 3030, 3040, 5030, 5040
The following patch adds support for the following Acer laptops to the
keyboard quirks:

Aspire: 1360, 1520, 1690, 3020, 3030, 3040, 5030, 5040
Ferrari: 4000

As per my discussion on the HAL mailing list with Richard Hughes, Fn+F3 (Power
Management) is remapped to KEY_BATTERY, and the extra Euro and Dollar keys
are left mapped to F14 and F15, but marked FIXME (pending a better solution
upstream).

-Carlos

(Small informative aside - many of Acer's models are identical versions of the
same hardware, just with a different processor. e.g. the 3020 and 5020 are
the same hardware, except with an AMD Sempron and Turion processor,
respectively).

1360 == 1520 (Odd, but the manual confirms this)
3020 == 5020
3030 == 5030
3040 == 5040