add keymap for Medion 9580-F
I attached yet another hal-info keymap patch, this time for an old
Medion 9580-F laptop. Please merge!

(I figured it might be a better idea to send this patch directly to
you instead of the ML, to lower the signal-to-noise ratio on the ML a
bit)

Oh, and the patch also updates the comments of my older MSI S270
keymap xml fragment.