quirk for Dell XPS M1210
Since the last set of fedora updates, the attached quirk is needed to
make a dell xps m1210 (with the integrated intel graphics card) suspend
reliably.

On iwl land, it now suspends and resumes without unloading the module
since fedora's 2.6.22.1-41 kernel. Though with the iwl patch included in
that kernel removing module does no longer works, so to the general
advise (on hal quirk site) to put iwl3945 in SUSPEND_MODULES, a note
should be added to try suspending without it first on recent kernels.
Otherwise adding the module will just put the system in a mostly useless
state when trying to suspend.