fix brigtness_in_hardware for Lenovo ThinkPads
Wondered about the matches/entries in 10-laptop-panel-hardware.fdi for each
different Lenovo ThinkPad machine. I have not see a Lenovo ThinkPad yet which
don't handle brightness in hardware.

I would propose to change the match for now to set the property
laptop_panel.brightness_in_hardware to true for all Lenovo ThinkPads. If
there are machines in the future which don't support this we can blacklist
them or add other generic matches.