swap errornous MMC and SmartMedia entries for card reader
From Carlos Corbacho <cathectic@slackadelic.com>:

I have a card reader here that matches the product & device ID of the "Noname
multi-card Reader" (this particular one is a "Vantec 7-in-1").

However, on my device, the LUNs for Smart Media and Memory Stick appear to be
reversed (patch below to illustrate - I don't have any CF or SD/ MMC cards,
so cannot vouch for those entries).

Is this a case of an error in the original patch, or do we have multiple card
readers with the same IDs, but different devices on different LUNs?