Suspend Patch for Dell Inspiron 1420n
Tested with Dell Inspiron 1420n pre-installed with Ubuntu.  On  a
clean Fedora 8 Test 3 install, suspend doesn't work.  Adding 1420
to the 'contains' match under inspiron results in a working suspend.
Still trying to get wireless (iwl3945) to survive through the suspend,
but that's a bit a easier when your screen comes back!