Dell D630 suspend patch
Please find attached patch needed to get the graphical mode restore to
work after suspend on Dell Latitude D630. I observe that suspend does
not work if I close the lid at the login screen, or in the text mode.
Otherwise, I have tested it a number of times, rebooting the laptops
few times in between.