fix Acer TravelMate 4650 keymap
A keymap for this laptop was added in commit 3d6b16, but is
clearly wrong - from the physical appearance at least, we are
missing the mapping for the Euro and Dollar keys at a minimum.

This laptop is more likely a "Laptops with extra Euro and Dollar keys, and
Fn+Right (Brightness Up) needs mapping (type 2)", so change it to that.