hp tx1000 hal-info keymap patch
Attached is a patch file for the HP tx1000 series. I have a tx1340ea but I think the keymap will work on most tx1000 laptops.

One problem I had with the keymap was the DVD and PC/PVR mappings. I took these from the sorted list and referred to the input.h on my machine (kernel 2.6.23). One button starts a DVD player and the Other button starts a media player in media mode. Not sure what I did wrong but they are still showing up in /var/log/messages as being unmapped.