identify modems and mobile broadband cards
As discussed in previous threads.  fdi patch to hal-info attached as
it's too big for most mailing lists.  If we find things in the
ETSI/UMTS/3GPP stuff that is _new_ to UMTS cards, we'll add another
command set specification.  There are plenty of EDGE/GPRS cards and
actual phones that don't support UMTS and therefore wouldn't have any
new stuff like that.