ported USB (mass storage) music players to new keys
Ported existing USB music players which are listed to be supported via
USB Mass Storage (UMS) to the new keys. These are the changes:

- portable_audio_player.access_method=storage was changed to:
portable_audio_player.access_method.protocols={storage} (strlist)
- portable_audio_player.type was removed

The old keys get readded for temporary backward compability at the end of
the file where we merge already common keys for storage devices. We need to
removed them if the the using applications are updated.

Note: we need to port also the keys for the player with
      portable_audio_player.access_method=user. Not sure atm which one is
      supported by which protocol, assume mtp. Maybe remove them completely
      so that the related lib/tool has to provide the info via a own
      FDI-file.