blacklist CD drive of the Huawei E220 USB HSDPA modem
The Huawei E220 is a USB HSPDA/GPRS modem. It appears to the
computer as a USB CD drive, on which are drivers for Windows
systems. Of course, we need no such drivers, so the CD drive
is useless and annoying as it is mounted by GNOME whenever a
user logs in.